fix full scan detection
When index scan direction is reversed, start_key and end_key are
reversed. This means that a full index scan can be detected by looking
for start_key="<MAX>", end_key=null|[].

This changes the detection code to use an order-independent comparison.
It concatenates start_key and end_key in an array, remove empty or
null elements, and test that what's left is a single element "<MAX>".